backend/requirements.txt
Django==3.1.4
django-cors-headers==3.6.0
django-environ==0.4.5
djangorestframework==3.12.2
djangorestframework-simplejwt
environ==1.0
GitPython==3.1.12
Jinja2
pip-tools==5.5.0
psycopg2-binary==2.8.6
PyGithub==1.54.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python-slugify==4.0.1
PyJWT[crypto]==1.7.1
PyYAML
requests
djangorestframework-jwt
pytest
